fix(test): consolidate duplicate tests, fix 6 @Disabled tests, eliminate false error logs
- Consolidate 3 Round 5/Round 9 duplicate test pairs:
  RestImportServiceDeepBranchTest → RestImportServiceUncoveredBranchTest
  A2AToolProviderManagerDeepBranchTest → A2AToolProviderManagerUncoveredBranchTest
  GroupConversationServiceDeepBranchTest → GroupConversationServiceUncoveredBranchTest
- Fix 6 @Disabled tests (3 StructuralMatcher + 3 GroupConversation):
  StructuralMatcher: short IDs → valid 24-char hex IDs for extractResourceId
  GroupConversation: stub jsonSerialization.serialize() for fallback path
- Replace http://example.com URLs with file:///nonexistent to avoid Swagger ERROR logs
- Replace unreachable DNS names with 127.0.0.1:1 for instant connection refused
  (also speeds up test suite by eliminating DNS timeout waits)
