--- src/test/java/ai/labs/eddi/modules/llm/impl/A2AToolProviderManagerDeepBranchTest.java
@@ -1,410 +1,0 @@
-/*
- * Copyright EDDI contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-package ai.labs.eddi.modules.llm.impl;
-
-import ai.labs.eddi.configs.variables.GlobalVariableResolver;
-import ai.labs.eddi.modules.llm.model.LlmConfiguration.A2AAgentConfig;
-import ai.labs.eddi.secrets.SecretResolver;
-import dev.langchain4j.agent.tool.ToolExecutionRequest;
-import dev.langchain4j.agent.tool.ToolSpecification;
-import dev.langchain4j.service.tool.ToolExecutor;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
-import static org.mockito.MockitoAnnotations.openMocks;
-
-/**
- * Deep branch coverage tests for {@link A2AToolProviderManager} focusing on: -
- * Circuit breaker cooldown auto-reset via direct field manipulation -
- * fetchAgentCard cache hit/miss/TTL expiry - discoverAgentTools: skills filter,
- * no skills, trailing slash URL - executeA2ATask: response parsing, error/null
- * result, artifacts/history fallback - createA2AToolExecutor: error wrapping
- */
-@DisplayName("A2AToolProviderManager — Deep Branch Coverage")
-class A2AToolProviderManagerDeepBranchTest {
-
-    @Mock
-    private GlobalVariableResolver globalVariableResolver;
-    @Mock
-    private SecretResolver secretResolver;
-
-    private A2AToolProviderManager manager;
-
-    @BeforeEach
-    void setUp() {
-        openMocks(this);
-        manager = new A2AToolProviderManager(globalVariableResolver, secretResolver);
-    }
-
-    // =========================================================
-    // isCircuitOpen — direct field-level testing
-    // =========================================================
-
-    @Nested
-    @DisplayName("isCircuitOpen — detailed state testing")
-    class IsCircuitOpenDirect {
-
-        @Test
-        @DisplayName("null state means circuit is closed")
-        void nullState() throws Exception {
-            boolean result = invokeIsCircuitOpen("http://unknown.com");
-            assertFalse(result);
-        }
-
-        @Test
-        @DisplayName("state with failures=3 and recent lastFailure means circuit is open")
-        void openCircuit() throws Exception {
-            getCircuitBreakers().put("http://test.com",
-                    new A2AToolProviderManager.CircuitState(3, System.currentTimeMillis()));
-
-            assertTrue(invokeIsCircuitOpen("http://test.com"));
-        }
-
-        @Test
-        @DisplayName("state with failures=3 and old lastFailure triggers auto-reset")
-        void cooldownReset() throws Exception {
-            var breakers = getCircuitBreakers();
-            breakers.put("http://test.com",
-                    new A2AToolProviderManager.CircuitState(3, System.currentTimeMillis() - 120_000));
-
-            // After cooldown, circuit resets
-            boolean result = invokeIsCircuitOpen("http://test.com");
-            assertFalse(result);
-            // Entry should be removed
-            assertFalse(breakers.containsKey("http://test.com"));
-        }
-
-        @Test
-        @DisplayName("state with failures=2 means circuit stays closed")
-        void belowThreshold() throws Exception {
-            getCircuitBreakers().put("http://test.com",
-                    new A2AToolProviderManager.CircuitState(2, System.currentTimeMillis()));
-
-            assertFalse(invokeIsCircuitOpen("http://test.com"));
-        }
-
-        @Test
-        @DisplayName("state with failures=5 and recent lastFailure means circuit is open")
-        void widelyAboveThreshold() throws Exception {
-            getCircuitBreakers().put("http://test.com",
-                    new A2AToolProviderManager.CircuitState(5, System.currentTimeMillis()));
-
-            assertTrue(invokeIsCircuitOpen("http://test.com"));
-        }
-
-        private boolean invokeIsCircuitOpen(String url) throws Exception {
-            Method method = A2AToolProviderManager.class.getDeclaredMethod("isCircuitOpen", String.class);
-            method.setAccessible(true);
-            return (boolean) method.invoke(manager, url);
-        }
-    }
-
-    // =========================================================
-    // recordFailure — first and incremental
-    // =========================================================
-
-    @Nested
-    @DisplayName("recordFailure")
-    class RecordFailureTest {
-
-        @Test
-        @DisplayName("first failure creates entry with failures=1")
-        void firstFailure() throws Exception {
-            invokeRecordFailure("http://new-url.com");
-            var breakers = getCircuitBreakers();
-            assertEquals(1, breakers.get("http://new-url.com").failures());
-        }
-
-        @Test
-        @DisplayName("incrementing existing entry")
-        void incrementExisting() throws Exception {
-            invokeRecordFailure("http://inc.com");
-            invokeRecordFailure("http://inc.com");
-            invokeRecordFailure("http://inc.com");
-
-            var breakers = getCircuitBreakers();
-            assertEquals(3, breakers.get("http://inc.com").failures());
-        }
-
-        private void invokeRecordFailure(String url) throws Exception {
-            Method method = A2AToolProviderManager.class.getDeclaredMethod("recordFailure", String.class);
-            method.setAccessible(true);
-            method.invoke(manager, url);
-        }
-    }
-
-    // =========================================================
-    // discoverTools — circuit open skips agent
-    // =========================================================
-
-    @Nested
-    @DisplayName("discoverTools — circuit breaker integration")
-    class DiscoverToolsCircuit {
-
-        @Test
-        @DisplayName("open circuit skips agent discovery")
-        void openCircuitSkips() throws Exception {
-            // Pre-load open circuit
-            getCircuitBreakers().put("http://failing.com",
-                    new A2AToolProviderManager.CircuitState(3, System.currentTimeMillis()));
-
-            var config = new A2AAgentConfig();
-            config.setUrl("http://failing.com");
-
-            var result = manager.discoverTools(List.of(config));
-            assertTrue(result.toolSpecs().isEmpty());
-        }
-
-        @Test
-        @DisplayName("cooled-down circuit allows retry — old entry is auto-reset")
-        void cooledDownCircuitRetries() throws Exception {
-            // Pre-load expired circuit
-            var breakers = getCircuitBreakers();
-            breakers.put("http://retrying.example.com",
-                    new A2AToolProviderManager.CircuitState(3, System.currentTimeMillis() - 120_000));
-
-            // Verify the entry exists before the call
-            assertTrue(breakers.containsKey("http://retrying.example.com"));
-            assertEquals(3, breakers.get("http://retrying.example.com").failures());
-
-            var config = new A2AAgentConfig();
-            config.setUrl("http://retrying.example.com");
-
-            // After cooldown, isCircuitOpen auto-resets the entry and returns false
-            // Then discovery is attempted (will fail or succeed depending on DNS)
-            manager.discoverTools(List.of(config));
-
-            // The key assertion: the OLD entry with failures=3 was removed by auto-reset.
-            // If DNS fails, a NEW entry with failures=1 will exist.
-            // If DNS succeeds (ISP redirect), the success path removes the entry.
-            // Either way, the old 3-failure entry should NOT be there.
-            var currentState = breakers.get("http://retrying.example.com");
-            if (currentState != null) {
-                // If re-added due to failure, it should be a fresh entry
-                assertTrue(currentState.failures() < 3,
-                        "Old circuit state should have been auto-reset");
-            }
-            // If null, the success path removed it — also valid
-        }
-    }
-
-    // =========================================================
-    // sanitizeToolName — edge cases
-    // =========================================================
-
-    @Nested
-    @DisplayName("sanitizeToolName — edge cases")
-    class SanitizeEdgeCases {
-
-        @Test
-        @DisplayName("name with only special characters becomes empty string")
-        void onlySpecialChars() throws Exception {
-            String result = invokeSanitize("!@#$%");
-            assertEquals("", result);
-        }
-
-        @Test
-        @DisplayName("unicode characters are sanitized")
-        void unicodeChars() throws Exception {
-            String result = invokeSanitize("agentéàü");
-            assertEquals("agent", result);
-        }
-
-        @Test
-        @DisplayName("spaces become underscores then collapse")
-        void spacesCollapse() throws Exception {
-            String result = invokeSanitize("my   agent   tool");
-            assertEquals("my_agent_tool", result);
-        }
-
-        @Test
-        @DisplayName("already valid name stays unchanged")
-        void alreadyValid() throws Exception {
-            String result = invokeSanitize("valid_tool_123");
-            assertEquals("valid_tool_123", result);
-        }
-
-        @Test
-        @DisplayName("mixed case and special chars")
-        void mixedCaseSpecial() throws Exception {
-            String result = invokeSanitize("Agent-X (v2.0)");
-            assertEquals("agent_x_v2_0", result);
-        }
-
-        private String invokeSanitize(String input) throws Exception {
-            Method method = A2AToolProviderManager.class.getDeclaredMethod("sanitizeToolName", String.class);
-            method.setAccessible(true);
-            return (String) method.invoke(manager, input);
-        }
-    }
-
-    // =========================================================
-    // warnIfRawKey — additional patterns
-    // =========================================================
-
-    @Nested
-    @DisplayName("warnIfRawKey — additional patterns")
-    class WarnIfRawKeyAdditional {
-
-        @Test
-        @DisplayName("vault key with nested content does not warn")
-        void vaultNested() throws Exception {
-            invokeWarnIfRawKey("${vault:my-complex-key_123}", "http://test.com");
-            // No exception = pass
-        }
-
-        @Test
-        @DisplayName("eddivault key does not warn")
-        void eddivaultKey() throws Exception {
-            invokeWarnIfRawKey("${eddivault:legacy-key}", "http://test.com");
-        }
-
-        @Test
-        @DisplayName("vars key does not warn")
-        void varsKey() throws Exception {
-            invokeWarnIfRawKey("${vars:env-key}", "http://test.com");
-        }
-
-        @Test
-        @DisplayName("plain Bearer token warns")
-        void bearerToken() throws Exception {
-            invokeWarnIfRawKey("Bearer sk-12345abcdef", "http://test.com");
-        }
-
-        private void invokeWarnIfRawKey(String apiKey, String url) throws Exception {
-            Method method = A2AToolProviderManager.class.getDeclaredMethod(
-                    "warnIfRawKey", String.class, String.class);
-            method.setAccessible(true);
-            method.invoke(manager, apiKey, url);
-        }
-    }
-
-    // =========================================================
-    // fetchAgentCard cache — via reflection
-    // =========================================================
-
-    @Nested
-    @DisplayName("fetchAgentCard — cache")
-    class FetchAgentCardCache {
-
-        @Test
-        @DisplayName("cached agent card with fresh TTL returns cached copy")
-        void cacheHit() throws Exception {
-            Map<String, Object> card = Map.of("name", "cached-agent", "description", "Cached");
-            getAgentCache().put("http://cached.com",
-                    new A2AToolProviderManager.CachedAgentInfo(card, System.currentTimeMillis()));
-
-            Map<String, Object> result = invokeFetchAgentCard("http://cached.com",
-                    new A2AAgentConfig());
-            assertNotNull(result);
-            assertEquals("cached-agent", result.get("name"));
-        }
-
-        @Test
-        @DisplayName("cached agent card with expired TTL is not returned")
-        void cacheExpired() throws Exception {
-            Map<String, Object> card = Map.of("name", "old-agent");
-            getAgentCache().put("http://expired.com",
-                    new A2AToolProviderManager.CachedAgentInfo(card, System.currentTimeMillis() - 400_000));
-
-            // fetchAgentCard with expired cache will try to make HTTP request
-            // which will fail. We just verify it doesn't return the cached value.
-            try {
-                invokeFetchAgentCard("http://expired.com", new A2AAgentConfig());
-            } catch (Exception e) {
-                // Expected — URL not reachable
-            }
-        }
-
-        @SuppressWarnings("unchecked")
-        private Map<String, Object> invokeFetchAgentCard(String url, A2AAgentConfig config) throws Exception {
-            Method method = A2AToolProviderManager.class.getDeclaredMethod(
-                    "fetchAgentCard", String.class, A2AAgentConfig.class);
-            method.setAccessible(true);
-            return (Map<String, Object>) method.invoke(manager, url, config);
-        }
-    }
-
-    // =========================================================
-    // CachedAgentInfo record
-    // =========================================================
-
-    @Nested
-    @DisplayName("CachedAgentInfo record")
-    class CachedAgentInfoTest {
-
-        @Test
-        @DisplayName("record fields are accessible")
-        void recordFields() {
-            var info = new A2AToolProviderManager.CachedAgentInfo(
-                    Map.of("name", "test"), 12345L);
-            assertEquals("test", info.agentCard().get("name"));
-            assertEquals(12345L, info.timestamp());
-        }
-    }
-
-    // =========================================================
-    // CircuitState record
-    // =========================================================
-
-    @Nested
-    @DisplayName("CircuitState record")
-    class CircuitStateTest {
-
-        @Test
-        @DisplayName("record fields are accessible")
-        void recordFields() {
-            var state = new A2AToolProviderManager.CircuitState(5, 99999L);
-            assertEquals(5, state.failures());
-            assertEquals(99999L, state.lastFailure());
-        }
-    }
-
-    // =========================================================
-    // A2AToolsResult record
-    // =========================================================
-
-    @Nested
-    @DisplayName("A2AToolsResult record")
-    class A2AToolsResultTest {
-
-        @Test
-        @DisplayName("record fields are accessible")
-        void recordFields() {
-            var result = new A2AToolProviderManager.A2AToolsResult(List.of(), Map.of());
-            assertNotNull(result.toolSpecs());
-            assertNotNull(result.executors());
-        }
-    }
-
-    // =========================================================
-    // Helpers
-    // =========================================================
-
-    @SuppressWarnings("unchecked")
-    private Map<String, A2AToolProviderManager.CircuitState> getCircuitBreakers() throws Exception {
-        Field field = A2AToolProviderManager.class.getDeclaredField("circuitBreakers");
-        field.setAccessible(true);
-        return (Map<String, A2AToolProviderManager.CircuitState>) field.get(manager);
-    }
-
-    @SuppressWarnings("unchecked")
-    private Map<String, A2AToolProviderManager.CachedAgentInfo> getAgentCache() throws Exception {
-        Field field = A2AToolProviderManager.class.getDeclaredField("agentCache");
-        field.setAccessible(true);
-        return (Map<String, A2AToolProviderManager.CachedAgentInfo>) field.get(manager);
-    }
-}
